+++ src/gpu/registers/power_control_register2.rs
@@ -1,5 +1,5 @@
 bitflags! {
-  pub struct PowerControlRegister2: u32 {
+  pub struct PowerControlRegister2: u16 {
     const SOUND_SPEAKERS_ENABLE = 1;
     const WIFI_ENABLE = 1 << 1;
   }
